requirements.txt
alembic==1.16.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
blinker==1.9.0
cffi==1.17.1
click==8.2.1
Flask==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
pycparser==2.22
python-dotenv==1.1.1
SQLAlchemy==2.0.42
typing_extensions==4.14.1
Werkzeug==3.1.3
WTForms==3.2.1